feat: reproducible bytecode CI, timelock boundary tests, trust model docs, fork tests
This PR addresses four issues:

113 — CI: add reproducible-bytecode verification job
- Add .github/workflows/reproducible-bytecode.yml that compiles with pinned
  solc 0.8.24 and records creation/runtime bytecode SHA256 hashes
- Pin initial hashes in contracts/evm/.pinned-bytecode/
- Document exact compiler settings (solc, EVM version, optimizer) and repro
  build instructions in docs/deployment.md

105 — Testing: timelock boundary tests
- Add test_RevertWhen_ExecuteOneSecondBeforeReadyAt: execute at readyAt - 1
  reverts NotReady (off-by-one)
- Add test_ExecuteExactlyAtReadyAt_Succeeds: execute at exactly readyAt
  succeeds (off-by-one)
- Add test_RevertWhen_RemoveOwnerCalledDirectly, test_RevertWhen_SetThresholdCalledDirectly,
  test_RevertWhen_SetDelayCalledDirectly: all onlySelf config setters revert
  NotSelf on direct call
- Add test_RemoveOwnerThroughGovernance, test_SetThresholdThroughGovernance:
  config changes succeed via fully-confirmed delayed ops
- Add event assertions (Proposed, Confirmed, Executed) to
  test_AddOwnerThroughGovernance
- Fix pre-existing  sequencing bug where MIN_DELAY/MAX_DELAY
  getters consumed the expected-revert before the constructor call
- Declare missing MIN_DELAY (1 hour) and MAX_DELAY (30 days) constants
- Declare missing GUARDIAN_PAUSE_TTL (72h), MAX_DESTINATION_LEN (56),
  MAX_DEST_ASSET_LEN (69) constants
- Fix pre-existing Released event signature mismatch in escrow test
- Fix pre-existing BadProtocolVersion test to expect UnknownVersion error
- Fix _decodeFillConfirmed to decode informational fillAmount/fillLedger fields

123 — Docs: document precise trust model
- Add consolidated trust model section to docs/threat-model.md as single
  source of truth: role-by-role (user, solver, relayer, LZ endpoint, DVN set,
  Stellar validators, admin/timelock, guardian, executor)
- Each role enumerated with trusted assumption, misbehavior if compromised,
  worst-case impact (steal/censor/delay/grief/halt), and bounding mechanism
- Key explicit assumptions: LZ DVN integrity, Stellar consensus, EVM finality,
  deterministic bytecode, EVM-address identity
- Cross-reference from docs/deployment.md trust model table

107 — Testing: add fork tests against real LZ endpoint / ERC-20s
- Add contracts/evm/test/Fork.t.sol with fork tests gated on FORK_RPC_URL
- Tests: lock with real USDC, lock with real USDT (no-return), fee-too-low
  on real endpoint, measured-delta with non-fee-on-transfer, endpoint auth
- Document how to run with FORK_RPC_URL environment variable

Closes #113, Closes #105, Closes #123, Closes #107

+++ contracts/evm/test/AmountBoundary.t.sol
@@ -45,7 +45,7 @@ contract AmountBoundaryTest is Test {
     // u128::MAX = 2^128 - 1
     uint256 internal constant U128_MAX = type(uint128).max;
     // i128::MAX = 2^127 - 1 = u128::MAX >> 1
-    uint256 internal constant I128_MAX = uint256(type(int128).max);
+    uint256 internal constant I128_MAX = uint256(uint128(type(int128).max));
 
     AmountBoundaryHarness internal harness;
